Time: 1 ms (99.36%), Space: 42.8 MB (24.36%) - LeetHub

+++ 3871-count-commas-in-range-ii/3871-count-commas-in-range-ii.java
@@ -1,8 +1,8 @@
 class Solution {
     public long countCommas(long n) {
         long ans = 0;
-        long length = String.valueOf(Math.abs(n)).length();
-        long multiple = 0;
+        int length = String.valueOf(Math.abs(n)).length();
+        int multiple = 0;
         if (length % 3 == 0) {
             multiple = (length / 3) - 1;
         } else {
@@ -11,7 +11,7 @@ class Solution {
         if (multiple >= 1) {
             long small = 1000;
             long large = 999999;
-            for (long i = 1; i <= multiple; i++) {
+            for (int i = 1; i <= multiple; i++) {
                 if (i != multiple) {
                     ans += (large - small + 1) * i;
                 } else {
